Sketch design for new open-in-editor feature
The idea is to add a second public method to the Appear library,
`Appear#open_file(filename, :editor => :tmux_nvim_ide)`
This would open the file in the named editor, and appear the process
containing the editor.

The sketch contains the design of a TmuxNvimIde, which edits a file in
an nvim in a nice tmux split.

Implementation will require a greatly expanded tmux library, with some
knowledge of tmux commands? Or I'll simplify the design and pass around
structs like a normal person.

We'll see!!