Lsof: ignore failure exit codes; always parse result
LSOF will exit non-zero in a bunch of cases where I'd expect it to
exit 0, like if it excludes a result based on your filtering parameters
or something. This lead to a bug where I couldn't reveal a process in
iTerm2, even though manually running `lsof` on the tty showed the
information we needed.