Tmux: improve targeting and return types
Tmux creation commands did not properly return the newly created
objects, because you must pass the -P flag to tmux to get info about
created objects, and I wasn't.

This commit corrects this with a new ipc method ipc_returning_one, which
passes the -P command for us, and also returns a single item instead of
a list.

Tmux commands like split_window now target panes and windows by ID
instead of indexes. This allows code that may change the index of a
conceptual "window" object to refer to the same window with a consistent
variable.