unix-dropper.applescript: tool to run a script on dropped files
This Applescript is a flexible tool that can run any shell script on
dropped files after saving this file as an Application in Script
Editor.app.

To configure the shell script that the app uses, just launch it via
double-click and edit the preference as necessary.

After you've configured the script, you can either drop files on the
.app, or set the .app as the default file handler for a filetype. Then,
double-click, and you script will run on the file!

![screenshot](https://camo.githubusercontent.com/c1e9ffd09b3d1f160880273a393d81c6a9a0f22d/68747470733a2f2f6d6f6e6f736e61702e636f6d2f66696c652f424d6d433337785137657133544a43687a7055506a70615a5055634533682e706e67)